+++ packages/zephyr/stories/Tooltip.stories.tsx
@@ -29,7 +29,8 @@ export const Default: Story<TooltipProps> = () => {
 
   return (
     <Tooltip
-      label="Now you see me!"
+      withArrow={true}
+      label={'Now you see me!'}
       data-testid={meta.title}
       side="right"
       manualControlProps={
@@ -192,6 +193,54 @@ export const WithoutBorder = () => (
     </Tooltip>
   </Box>
 );
+
+export const WithCustomStyles = () => {
+  return (
+    <Tooltip
+      tx={{
+        TooltipArrow: {
+          fill: '#B4CFFF',
+          stroke: '#B4CFFF',
+        },
+        TooltipContent: {
+          color: '#0B1953',
+        },
+        bg: '#B4CFFF',
+      }}
+      withArrow={true}
+      withBorder={false}
+      label={'Custom is cool'}
+      data-testid={meta.title}
+      side="bottom"
+      sideOffset={3}
+      manualControlProps={
+        isChromatic() ? { open: true, onOpenChange: noop, defaultOpen: true } : undefined
+      }
+    >
+      <Button onClick={noop} variant="button-filled-blue">
+        Hover to see a tooltip with custom styles
+      </Button>
+    </Tooltip>
+  );
+};
+
+export const WithoutArrow = () => {
+  return (
+    <Tooltip
+      withArrow={false}
+      label={'Arrows are boring, be a rebel'}
+      data-testid={meta.title}
+      side="right"
+      manualControlProps={
+        isChromatic() ? { open: true, onOpenChange: noop, defaultOpen: true } : undefined
+      }
+    >
+      <Button onClick={noop} variant="button-filled-blue">
+        Hover over me to see a tooltip without an arrow
+      </Button>
+    </Tooltip>
+  );
+};
 
 export const MultiColoredBackground = () => {
   const avatarSize = 36;